female ahora es 1 para mujer 0 para hombre. Close #4

+++ getEPH.py
@@ -233,7 +233,7 @@ def schoolYears(dataset):
 
 
 def categorize(df):
-    df.female = df.female == 2
+    df.female = (df.female == 2).astype(int)
     df.schoolLevel.replace(to_replace=[99], value=[np.nan] , inplace=True, axis=None) 
     df.lastYear.replace(to_replace=[98,99], value=[np.nan, np.nan] , inplace=True, axis=None)
     df.activity.replace(to_replace=[0], value=[np.nan] , inplace=True, axis=None)
